requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
psycopg2
python-dotenv
bcrypt
pydantic[email]
watchdog==3.0.0
python-jose[cryptography]
python-multipart
